+++ src/main/java/com/example/TaskSpringBoot1/services/SovereignService.java
@@ -1,6 +1,9 @@
 package com.example.TaskSpringBoot1.services;
 
+import com.example.TaskSpringBoot1.entity.Planet;
 import com.example.TaskSpringBoot1.entity.Sovereign;
+import com.example.TaskSpringBoot1.exception.SovereignException;
+import com.example.TaskSpringBoot1.repository.PlanetRepository;
 import com.example.TaskSpringBoot1.repository.SovereignRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -12,10 +15,12 @@ import java.util.stream.Collectors;
 @Service
 public class SovereignService {
     SovereignRepository sovereignRepository;
+    PlanetRepository planetRepository;
 
     @Autowired
-    public SovereignService(SovereignRepository sovereignRepository) {
+    public SovereignService(SovereignRepository sovereignRepository, PlanetRepository planetRepository) {
         this.sovereignRepository = sovereignRepository;
+        this.planetRepository = planetRepository;
     }
 
     public List<String> rankingByAge() {
@@ -24,8 +29,12 @@ public class SovereignService {
         return nameList;
     }
 
-    public void delete(long id) {
+    public void deleteById(long id) {
+        if (!sovereignRepository.existsById(id)) throw new SovereignException(" Recording does not exist ");
         sovereignRepository.deleteById(id);
+        List<Planet> planetList = planetRepository.getListBySovereign(id);
+        List<Planet> planetList1 = planetList.stream().peek(s -> s.setSovereign(null)).collect(Collectors.toList());
+        planetRepository.saveAll(planetList1);
     }
 
     public List<Sovereign> getListSovereign() {
